Update comment in example unicorn config
The timeout was changed to 30 and the comment wasn't updated. Also I've seen several reports of this timeout not being long enough for quite a few people. It resulted in me getting a 502 bad gateway from nginx until I upped it. Perhaps could back to the default 60 would be better? I changed mine to 300, running on an AWS micro instance.