Fix deletion of tmp/gitlabhq-gitolite* folders in a NFS environment.
When working inside of a Vagrant box with NFS enabled, the deletion of a
repository in tmp/ doesn't work every time.

It is related to NFS inability to delete a folder if it's still used by
a resource (e.g. `rm -rf ./folder` would leave the folder and a .nfs*
    file in it). In this case it's the temporary repository which can't
be deleted because `ga_repo` is still using it.

De-allocating ga_repo is not possible (thanks Ruby), but deleting the
folder it points to in the first place fixes the issue.

Reference:
http://stackoverflow.com/questions/11228079/python-remove-directory-error-file-exists