Add feature to support email-less omniauth backends
If backend do not provide auth.info.email, We generate it from auth.info.nickname and email_domain option passed to omniauth provider (if it's passed).

You should be able to pass this option to all allready implemented omniauth providers and this commit fixes it for ldap.

This feature is usefull in environments wherre login is done by LDAP, pam, shibboleth or similar and that backend is not providing email. (ldap is missing email field. pam can't normally know it. etc..)