Add namespace_id to project creation for a user in API
This patch just adds the possibility to change the namespace of a
project created for a user via the API.

It might be interesting, for an administrator, to create projects in
the global namespace for specific users. This is possible when we create a
project without specifying the owner so why not here.