Fixed a nginx configuration
For some reason, if the nginx configuration is bound to an IP, it breaks the install, causing all git clone/push/pull operations to die with a 'fatal: The remote end hung up unexpectedly' message. Changing this configuration seems to fix the problem. I reckon that this is just a temporary fix. I don't know the ins and outs of GitLab well enough to vouch for this change, but I do know that these instructions are broken, and that this update will fix them.