Adding support for customized avatar services.
This new feature allows the user avatars to be fetched from any other service, based on any of the users details, such as the username.
The lookup is performed by email, to load the user record, and then a given field is extracted from the user record to be substituted on the avatar image URL format string.