Updated command to perform backup...
When you run the backup raketask, the rake suggests that you shouldn't be running the job as root, but only after it creates a file structure owned by root that then later has to be fixed before running the task as the git user. Adding the sudo prefix to the command saves the hassle.