Ensuring create/delete branch/tag supports periods in the url.
This fixes an issue where periods could not be in the API url
(eg: /projects/1/repository/tags/v.1.0/HEAD)

Also - The api methods now just return that the API call was received
(Still cannot actually return branch information as that is done in a
worker process). This change was made mainly due to the Ruby API wrapper
expecting a JSON response; however it makes sense to return *something*
rather than just an empty response.

Documentation updated for this.