Reenable the create project button after errors
Currently if you try to create a project from an existing repo but give it a bad
URL, you see the "Import url should be a valid url" error and the form, but the
submit button is disabled meaning you can't fix the URL and resubmit the form.
This fix just enables the button when the project has an error.

Closes #4070