Swap placeholders of project compare view input fileds
Now, placeholders of compare index view like /my_project/compare are
shown as from 'master' to 'aa8b4ef'. Because of that user thinks that
the order of these input fields are expected TO and FROM.
But these fields are expected FROM and TO.
This commit fixes the misleading placeholder.