Clear redis cache and precompile assets during update to 6.0. Fix #4874 #4840
As a consequence on the namespaces change from 5.4 to 6.0, after upgrade
lots of links on the dashboard were not working (pointing to projects
without the group part on the url).
Also, after upgrading to 6.0-stable, the network (branch history) function
does not work. The progress bar rotates forever, but the network graph cannot
be generated.

Clearing redis cache fixes those issues.