Issue-4769 - Unable to update case of label
Labels are saved in the database with the case they were originally created with.
Before this change if a user created a label with the same text but different case
the label would use the original case and no new label would be created in the
database. With this change, labels are now case-sensitive.

Steps to test:

1. Before this change, create a new issue with a new label "FIxMe"
2. Edit the issue and change the label to "FixMe"
3. Note that the label reverted to "FIxMe"
4. Apply this change
5. Edit the issue again and change the label to "FixMe"
Note that the new case was preserved. If you also look in the database in the "tags" table you will see that both labels are present - "FIxMe" and "FixMe".