Remove instructions about setting vim as the default editor
The default editor of Debian/Ubuntu is nano because it's easy to use for beginners. If a user proficient in vim reads this guide he will either have set vim as his default editor or understand to change editor to vim.

Even if the vim user doesn't understand to make these changes nano is easy to use and will be of no issue to him. Vim on the other hand will confuse the user who accidentally follows the instructions which are entirely unrelated to an already long set of installation instructions and sets vim as editor.