Check if a file is editied "on" a branch or a commit.
The previous behavior was a "disabled" link, which was still clickable
and linked to the edit file action.

If the 	user accesses a path like "/edit/master/README" show the file
edit form (just like before).

If the user accesses a path like
"/edit/8d5c1f375ce99e5df84e26b0eafbf1448c6a001e/README", redirect the
user to the file detail page with a note, that he cannot edit a file
"on" a commit.

fixes #5482