Include public projects in project listings.
Firstly, this fixes the group page that would 404 even when the user had
public access to some of the projects in a group. This fixes issue #5553.
Public projects are now included in the "All" section of the projects
page. The default dashboard doesn't show public projects though since
that would cause too much of an overload of information if the number of
public projects is very large. We want to make it possible for them to
find public projects easily, but we don't want to overload them with
information in their daily workflow.

Slightly tweaked the message when a user doesn't have access to any
projects on the dashboard. It now includes a link to the public section
of the site so it's more obvious where to look for that information.

The search has also been improved to include public projects.