Changed sudo by su -c
The sudo command needs a tty to run. If you put this file on a CentOS 6, for example, it will not run in the host startup process. Instead, the 'su -c' runs the command apart if it has or not a tty.

I tested this change in a fresh install of gitlab on a CentOS 6.5 with successful.