Instruct users to fetch merge request branch
Instructing users to create a new branch on the target branch and then
pulling creates a few issues. If the target branch has moved on since
the source branch diverged from it, then the pull will create an
unnecessary merge commit from the target branch to the source branch. If
the user has pull.rebase set to "true" or "preserve", then this creates
an even stranger history. These instructions will ensure the local
branch created for the merge request is exactly what contributing user
has pushed.