Don't write the project namespace in email subjects
For instance, the email "Subject" field changes from:

    Mattt / Ground Control | Saving issue doesn't work sometimes (#1234)

to:

    Ground Control | Saving issue doesn't work sometimes (#1234)

Rationale:

* Most people are receiving email notifications about a single fork:
the project activity, issues and merge requests are happening either
in a root repository or in a fork, but less often in several forks.

* It removes noise from the email subject. The namespace of the project
can still be read in the email body.

* For Entreprise users that have a single namespace "MyEntreprise",
having this repeated in every email subject is very noisy.