Add spinner in favicon
Sometimes the site header is out of view, so when the user clicks some
long-loading link (e.g. large diff page), and the spinner in the header
starts spinning, user doesn't see it. So basically nothing happens for
them, which is confusing (is website working at all?).

This makes the favicon being replaced by a spinner (in user agents that
support this) when the header spinner is shown, and default favicon
restored when the header spinner is hidden.