fix link to notes in merge request diff
http://localhost:3000/gitlabhq/gitlab-ci/merge_requests/1/diffs#note_316
scroll in such links doesn't work. It happens because  on page there are 2 elements with id="note_316" in changes and discussion.
Browser takes first for scroll.
This solution works because changes loads without diffs so tag with id="note_316" just one.
Diffs loads with discussions, but in this case diff tag would be first.
It is dirty but working solution.