Import the database as the `git` user
This ensures that all tables created during the import belong to `git`.
If you import as a different user, such as the `postgres` superuser, you
may encounter issues where the GitLab database user cannot access tables
in gitlabhq_production, _even if_ `git` is the owner of
gitlabhq_production at the time of import.