Show build status badge in public project list
We use GitLab for a while now and often discover that the list of projects is a bit overwhelming for a lot of our developers. They might not know if it's safe to use that particular package or not.

Clicking through every project isn't sufficient for us, so a possibility to see the build status right from the list would be a nice improvement.

Cheers and keep up the good work!