Fixed issue with nginx config
With default_server set in the config file, nginx fails to start because it already has a default server. The suggested fix found somewhere is to delete the other default config (sites-enabled/default), but this makes any subdomain (instead of just git.example.com) point to GitLab. This fixes the issue. Note that if GitLab is the only server config setup that it should have default_server set.