Less verbose HipChat integration
Whenever a push with a lot of merges is made, the HipChat channel gets
overwhelmed. It is easy to loose everything else in all the GitLab
messages. It makes more sense to limit the number of commits shown.

The same problem happens with longer commit messages (like this one). It
will be useful to show only the subject of the commit message, not the
full one.