Fix pasting images into browser
This patch adds a binding to the paste event resulting in handlePaste()
function beeing called for image data (or text) paste processing.
JQuery selector is used to bind the paste event in order to make it work
for all html elements of css class textarea.markdown-area.

This way it is possible to paste images directly into the mardown editor
at least when using a browser fully supporting clipboardData (like
Google Chrome does).