Specify IPv6 only listener on NGINX
By default on Linux the IPv6 listener in NGINX also starts an IPv4 listener, unless explicitly told to not do so (which is what we are doing here with the ipv6only=on).

Failure to do so will result on an "address already binded" error which will prevent the server from starting.