Fix broadcast message to show up properly with new UI
With the new UI, the broadcast message is the first level child element
of the body tag and then render full width without respecting the width
of the left sidebar. This makes the message goes under the left sidebar
in smaller screen.

This commit fixes the issue by moving the message element into the
"page-with-sidebar" div so it will always render together with the main
content area with same look as before. The rendering for the search,
login and other view without left sidebar remains untouched.

Releases: master, 7-7-stable
Fixes: #1019