Toggle sidebar button more obvious
The toggle is now at the top of the sidebar because it is not noticeable
near the bottom. By placing it at the top, users will immediately know that
they can have more space if they desire versus on the bottom, they will have
to search for it and that's not desired.

Fixes #2044