Merge branch 'conditional-delete-taggings-indices' into 'master'
Conditionally remove indices that may not exist in pre-GitLab v6.7 installations

In pre-GitLab v6.7 installations, the `tagging` indices were directly added to the init schema and not added in a migration. Anyone using this older DB schema may not have these indicies, so only remove them if they are there.

Closes #1593

See merge request !626