Use .md as extention for wiki pages
**What does this do?**
It makes sure that when you create a wiki page via the web interface, the
extention is .md instead of .markdown

**Why is this needed?**
When you're using Gollum locally, it will create pages with the .md extention.
Also .md is the best known extention for markdown. This fix will make sure that
if you're using gollum or the webinterface, the extention will be the same.

**What issues does this fix?**
Fixes https://github.com/gitlabhq/gitlabhq/issues/5204

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>