Add satellite create command for omnibus installs
In at least the 7.9.1 omnibus release and after I was able to use
the `gitlab-rake` command to rebuild the satellites, rather than
using bundle directly. This doc entry brings the satellite section
into alignment with the other sections in the doc which show both
the use of gitlab-rake and the direct execution of the rake task
via bundle.

ci-skip