Merge branch 'fix/oauth_user_ldap_person' into 'master'
Fix behavior of ldap_person method in Gitlab::OAuth::User

Code tweaks in 45e9150a caused the ldap_person method to not return expected results.
Improved tests to cover the ldap_person method, which was previously stubbed.

Restored the previous implementation of ldap_person, as I could not find a more concise to write it that still works with all the test cases.

See merge request !837