Merge branch 'animated-logo' into 'master'
Animate the logo on hover

The logo is now rendered as pure SVG, rather than image referencing a
svg. The SVG has an id and the shapes of the logo have a shared class.
The shapes change their fill color on hover with a transition.

Identical to how it looks on about.gitlab.com

cc @skyruler

See merge request !1557